--- src/main/java/com/github/nathandekeyrel/kismet/Gender.java
@@ -1,6 +1,0 @@
-package com.github.nathandekeyrel.kismet;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
