+++ src/main/java/com/github/nathandekeyrel/kismet/profile/ProfileAnswer.java
@@ -1,5 +1,6 @@
-package com.github.nathandekeyrel.kismet;
+package com.github.nathandekeyrel.kismet.profile;
 
+import com.github.nathandekeyrel.kismet.user.User;
 import jakarta.persistence.*;
 import lombok.Data;
