+++ src/main/java/com/github/nathandekeyrel/kismet/user/Gender.java
@@ -1,0 +1,6 @@
+package com.github.nathandekeyrel.kismet.user;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
