+++ src/main/java/com/github/nathandekeyrel/kismet/user/User.java
@@ -1,6 +1,7 @@
-package com.github.nathandekeyrel.kismet;
+package com.github.nathandekeyrel.kismet.user;
 
 
+import com.github.nathandekeyrel.kismet.friendship.Friendship;
 import jakarta.persistence.*;
 import lombok.Data;
 
@@ -12,7 +13,6 @@ import java.util.Set;
 @Entity
 @Table(name = "users")
 public class User {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
